# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Employee Performance Rating API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      security:
        - bearerAuth: []
      parameters:
        - name: teamId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserProfile"
  /users/{userId}:
    get:
      operationId: getUser
      tags: [users]
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
    patch:
      operationId: updateUser
      tags: [users]
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserRequest"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
  /teams:
    get:
      operationId: listTeams
      tags: [teams]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of teams
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Team"
  /teams/{teamId}:
    get:
      operationId: getTeam
      tags: [teams]
      security:
        - bearerAuth: []
      parameters:
        - name: teamId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Team details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Team"
    patch:
      operationId: updateTeam
      tags: [teams]
      security:
        - bearerAuth: []
      parameters:
        - name: teamId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTeamRequest"
      responses:
        "200":
          description: Updated team
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Team"
  /users/register:
    post:
      operationId: registerUser
      tags: [users]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterUserRequest"
      responses:
        "201":
          description: User registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required: [error]
...
    UserProfile:
      type: object
      properties:
        userId:
          type: string
        displayName:
          type: string
        username:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [User, "Team Lead", Manager]
        level:
          type: string
          enum: [L1, L2, L3]
        teamId:
          type: integer
          nullable: true
        teamName:
          type: string
          nullable: true
      required: [userId, displayName, username, email, role, level]
    UpdateUserRequest:
      type: object
      properties:
        teamId:
          type: integer
          nullable: true
        role:
          type: string
    Team:
      type: object
      properties:
        teamId:
          type: integer
        teamName:
          type: string
        tlUserId:
          type: string
          nullable: true
        managerUserId:
          type: string
          nullable: true
        tlDisplayName:
          type: string
          nullable: true
        managerDisplayName:
          type: string
          nullable: true
      required: [teamId, teamName]
    UpdateTeamRequest:
      type: object
      properties:
        teamName:
          type: string
        tlUserId:
          type: string
          nullable: true
        managerUserId:
          type: string
          nullable: true
    RegisterUserRequest:
      type: object
      properties:
        displayName:
          type: string
        username:
          type: string
        email:
          type: string
        password:
          type: string
        role:
          type: string
        level:
          type: string
        teamId:
          type: integer
          nullable: true
      required: [displayName, username, email, password]
...